+++ src/main/java/com/example/community_cr/community/controller/dto/response/PostDetailResponse.java
