+++ src/main/java/com/example/community_cr/image/service/ImageService.java
@@ -1,0 +1,9 @@
+package com.example.community_cr.image.service;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface ImageService {
+	String upload(MultipartFile file);
+
+	String generatePresignedUrl(String fileName);
+}
